requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
mistralai>=1.0.0
sentence-transformers==3.0.1
faiss-cpu==1.8.0.post1
pypdf==5.0.1
numpy==2.1.1
